Simulation improvements
* in-band SNR uses EBW
* MSK now skips SRRC
* Fractional OSR via integer up, down
* Enforced up >= 2 for SRRC and up/down >= 1

+++ src/modreczoo/evaluation.py
@@ -161,7 +161,7 @@ def accuracy_by_osr_snr_levels(
                 {
                     "snr_bin_db": float(bin_start),
                     "snr_bin_end_db": float(bin_start + bin_width),
-                    "osr": int(value),
+                    "osr": float(value),
                     "n": int(np.sum(mask)),
                     "accuracy": float(accuracy_score(y_true[mask], y_pred[mask])),
                 }
